% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidSingleCell-package}
\alias{AcidSingleCell}
\alias{AcidSingleCell-package}
\title{AcidSingleCell}
\description{
Toolkit for single-cell RNA-seq analysis that extends the functionality
of SingleCellExperiment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidsinglecell/}
  \item \url{https://github.com/acidgenomics/r-acidsinglecell/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidsinglecell/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
