% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\name{AcidSingleCellTestsUrl}
\alias{AcidSingleCellTestsUrl}
\title{AcidSingleCell test data URL}
\usage{
AcidSingleCellTestsUrl
}
\description{
AcidSingleCell test data URL
}
\note{
Updated 2023-10-05.
}
\examples{
AcidSingleCellTestsUrl
}
\keyword{internal}
