% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregate-methods.R
\name{aggregate}
\alias{aggregate}
\alias{aggregate,SingleCellExperiment-method}
\title{Aggregate}
\usage{
aggregate(x, ...)

\S4method{aggregate}{SingleCellExperiment}(x, col = "aggregate", fun = "sum", MARGIN = 1L)
}
\arguments{
\item{x}{Object.}

\item{col}{\code{character(1)}.
Name of column in either \code{rowData()} or \code{colData()} that defines the
desired aggregation groupings.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{match.arg()} internally.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{...}{Additional arguments.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Aggregate
}
\note{
Updated 2021-09-13.
}
\section{Methods (by class)}{

\itemize{
\item \code{SummarizedExperiment}: Arguments pass through to \code{matrix} or \code{Matrix} method, depending on the class
of matrix defined in requested \code{assay}.
}
}

\section{Methods (by class)}{


\itemize{
\item \code{matrix}, \code{Matrix}:
Aggregate using a grouping \code{factor}.
\item \code{SummarizedExperiment}:
Aggregate data slotted in \code{assays()} using an automatically generated
grouping \code{factor}, which is obtained from a user-defined column
(\code{col} argument) in either the \code{rowData()} or \code{colData()} of the object.
Slot an \code{aggregate} column into \code{rowData()} for \code{aggregateRows()}, or into
\code{colData()} for \code{aggregateCols()}. This method will define the \code{groupings}
automatically, and perform the aggregation.
}

}

\examples{
data(SingleCellExperiment_lanesplit, package = "AcidTest")

## SingleCellExperiment ====
x <- SingleCellExperiment_lanesplit
levels(colData(x)[["aggregate"]])
x <- aggregate(
    x = x,
    col = "aggregate",
    fun = "sum",
    MARGIN = 2L
)
print(x)
}
\seealso{
\itemize{
\item \code{stats::aggregate()}.
\item \code{S4Vectors::aggregate()}.
\item \code{Matrix.utils::aggregate.Matrix()}.
\item \code{muscat::aggregateData()}.
}
}
\author{
Michael Steinbaugh, Rory Kirchner
}
