% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/aggregateCellsToSamples-methods.R
\name{aggregateCellsToSamples}
\alias{aggregateCellsToSamples}
\alias{aggregateCellsToSamples,SingleCellExperiment-method}
\title{Aggregate cells to samples}
\usage{
aggregateCellsToSamples(x, ...)

\S4method{aggregateCellsToSamples}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SummarizedExperiment}.
Object with cell-level counts aggregated to sample-level.
}
\description{
Utilty function that factilites cell-to-sample aggregation. By default, this
function will sum the counts across cells to sample level.
}
\details{
Internally this function automatically obtains the cell-to-sample groupings
and then performs a sum aggregation with the \code{aggregate()} function.
}
\note{
Updated 2021-09-13.
}
\examples{
data(SingleCellExperiment_lanesplit, package = "AcidTest")

## SingleCellExperiment ====
x <- SingleCellExperiment_lanesplit
x <- aggregateCellsToSamples(x)
print(x)
}
