% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateCols-methods.R
\name{aggregateCols}
\alias{aggregateCols}
\alias{aggregateCols,SingleCellExperiment-method}
\title{Aggregate columns}
\usage{
aggregateCols(x, ...)

\S4method{aggregateCols}{SingleCellExperiment}(x, fun)
}
\arguments{
\item{x}{Object.}

\item{fun}{\code{function}.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with columns aggregated.
}
\description{
Aggregate columns
}
\note{
Updated 2021-02-05.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")
sce <- SingleCellExperiment

## SingleCellExperiment ====
x <- cell2sample(sce)
table(x)
}
