% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cell2sample-methods.R
\name{cell2sample}
\alias{cell2sample}
\alias{cell2sample,SingleCellExperiment-method}
\title{Cell-to-sample mappings}
\usage{
cell2sample(object, ...)

\S4method{cell2sample}{SingleCellExperiment}(object, return = c("factor", "DFrame"))
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{"factor"}: Named \code{factor} containing the samples as the levels and cell
identifiers as the names.
\item \code{"DFrame"}: S4 data frame containing \code{sampleId} column and cell identifiers
as the row names.
}
}
\description{
Cell-to-sample mappings
}
\details{
Sample identifiers must be defined in \code{sampleId} column in \code{colData()}.
}
\note{
Updated 2022-05-04.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
c2s <- cell2sample(object)
table(c2s)
}
