% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellCountsPerCluster-methods.R
\name{cellCountsPerCluster}
\alias{cellCountsPerCluster}
\alias{cellCountsPerCluster,SingleCellExperiment-method}
\title{Cell counts per cluster}
\usage{
cellCountsPerCluster(object, ...)

\S4method{cellCountsPerCluster}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Cell counts per cluster
}
\note{
Updated 2021-09-14.
}
\examples{
data(SingleCellExperiment_Seurat, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_Seurat
x <- cellCountsPerCluster(object)
print(x)
}
