% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellTypesPerCluster-methods.R
\name{cellTypesPerCluster}
\alias{cellTypesPerCluster}
\alias{cellTypesPerCluster,KnownMarkers-method}
\title{Cell types per cluster}
\usage{
cellTypesPerCluster(object, ...)

\S4method{cellTypesPerCluster}{KnownMarkers}(object, min = 1L, max = Inf)
}
\arguments{
\item{object}{Object.}

\item{min, max}{\code{integer(1)}.
Minimum or maximum number of marker genes per cluster.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame}.
}
\description{
Cell types per cluster
}
\note{
Updated 2023-04-27.
}
\examples{
data(KnownMarkers, package = "AcidTest")

## KnownMarkers ====
object <- KnownMarkers
x <- cellTypesPerCluster(object)
print(x)
}
