% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combine multiple objects}
\usage{
combine(x, y, ...)

\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Combine multiple objects
}
\details{
We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges} and/or
\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}
\link[S4Vectors:Annotated-class]{metadata} must be identical.
\item \link[SummarizedExperiment:SummarizedExperiment-class]{colData} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} argument).
}
}
\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")
str_pad <- stringr::str_pad

## SingleCellExperiment ====
x <- SingleCellExperiment_splatter
colnames(x) <- paste0(
    "cell",
    str_pad(
        string = seq_len(ncol(x)),
        width = 4L,
        pad = "0"
    )
)

y <- x
colnames(y) <- paste0(
    "cell",
    str_pad(
        string = seq_len(ncol(y)) + ncol(y),
        width = 4L,
        pad = "0"
    )
)

## Combine the two objects.
c <- combine(x, y)
sampleData(c)
print(c)
}
\seealso{
\itemize{
\item \code{help("merge.Matrix", "Matrix.utils")}.
}
}
