% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertSampleIDsToNames-methods.R
\name{convertSampleIDsToNames}
\alias{convertSampleIDsToNames}
\alias{convertSampleIDsToNames,SingleCellExperiment-method}
\title{Convert sample identifiers to names}
\usage{
convertSampleIDsToNames(object, ...)

\S4method{convertSampleIDsToNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with valid sample IDs remapped to
human-readable sample names.
}
\description{
Convert sample identifiers to names
}
\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
convertSampleIDsToNames(object)
}
