% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,SingleCellExperiment-method}
\title{Export}
\usage{
export(object, ...)

\S4method{export}{SingleCellExperiment}(
  object,
  name = NULL,
  dir = getOption("acid.export.dir", default = "."),
  compress = getOption("acid.export.compress", default = FALSE),
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)} or \code{character(1)}.
These character strings are currently allowed for \code{\link[base:save]{save()}}:
\code{"gzip"}, \code{"bzip2"}, or \code{"xz"}.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisibily return file paths, either as \code{list} or \code{character}.
}
\description{
Export data out of R and write to disk.
}
\details{
This method extends \code{SummarizedExperiment} but also handles export of the
\code{reducedDims} slot (i.e. containing UMAP, tSNE, and PCA).
}
\note{
Updated 2021-02-06.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment
dir <- "example"
x <- export(object = object, dir = dir)
print(x)
unlink(dir, recursive = TRUE)
}
