% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geometricMean-methods.R
\name{geometricMean}
\alias{geometricMean}
\alias{geometricMean,sparseMatrix-method}
\title{Geometric mean}
\usage{
geometricMean(x, ...)

\S4method{geometricMean}{sparseMatrix}(x, MARGIN = 2L)
}
\arguments{
\item{x}{Object.}

\item{MARGIN}{\code{integer(1-2)}.
Dimension where the function will be applied.
For a two-dimensional matrix:
\code{1} indicates rows;
\code{2} indicates columns;
\code{c(1, 2)} indicates rows and columns.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
}
\description{
The geometric mean is the nth root of n products or e to the mean log of \code{x}.
}
\details{
Useful for describing non-normal (i.e. geometric) distributions.
}
\note{
Updated 2022-10-24.
}
\examples{
data(sparseMatrix, package = "AcidTest")

## Matrix ====
x <- sparseMatrix
geometricMean(x)
}
