% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{km}
\alias{km}
\title{Known markers example}
\format{
An object of class \code{KnownMarkers} with 10 rows and 10 columns.
}
\usage{
data(km)
}
\value{
\code{KnownMarkers}.
}
\description{
Known markers example
}
\note{
Updated 2021-09-14.
}
\examples{
data(km)
print(km)
}
\keyword{datasets}
