% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSingleCellExperiment.R
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\title{Make a SingleCellExperiment object}
\usage{
makeSingleCellExperiment(
  assays = S4Vectors::SimpleList(),
  rowRanges = GenomicRanges::GRanges(),
  colData = S4Vectors::DataFrame(),
  metadata = list(),
  reducedDims = S4Vectors::SSimpleList(),
  transgeneNames = NULL
)
}
\arguments{
\item{assays}{\code{SimpleList}.
Count matrices, which must have matching dimensions. Counts can be passed
in as either a dense matrix (\code{matrix}) or sparse matrix (\code{sparseMatrix}).}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{colData}{\code{DataFrame}.
Metadata describing the assay columns. For bulk RNA-seq, this data
describes the samples. For single-cell RNA-seq, this data describes the
cells.}

\item{metadata}{\code{list}.
Metadata.}

\item{reducedDims}{\code{SimpleList}.
List containing matrices of cell coordinates in reduced space.}

\item{transgeneNames}{\code{character}.
Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Make a SingleCellExperiment object
}
\note{
Updated 2021-02-05.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SimpleList ====
object <- SingleCellExperiment
assays <- assays(object)
rowRanges <- rowRanges(object)
colData <- colData(object)
metadata <- metadata(object)
reducedDims <- reducedDims(object)

x <- makeSingleCellExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    metadata = metadata,
    reducedDims = reducedDims
)
print(x)
}
\seealso{
\itemize{
\item \code{AcidExperiment::makeSummarizedExperiment()}.
\item \code{SingleCellExperiment::altExps()}, which has replaced the now defunct
\code{isSpike} method for setting spike-in transcripts.
}
}
