% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/melt-methods.R
\name{melt}
\alias{melt}
\alias{melt,SingleCellExperiment-method}
\title{Melt data from wide to long format}
\usage{
melt(object, ...)

\S4method{melt}{SingleCellExperiment}(
  object,
  assay = 1L,
  min = -Inf,
  minMethod = c("absolute", "perRow"),
  trans = c("identity", "log2", "log10")
)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{min}{\code{numeric(1)} or \code{NULL}.
Minimum count threshold to apply. Filters using "greater than or equal to"
logic internally. Note that this threshold gets applied prior to
logarithmic transformation, when \code{trans} argument applies.
Use \code{-Inf} or \code{NULL} to disable.}

\item{minMethod}{\code{character(1)}.
Only applies when \code{min} argument is numeric.
Uses \code{match.arg()}.
\itemize{
\item \code{absolute}: Applies hard cutoff to \code{counts} column after the melt
operation. This applies to all counts, not per feature.
\item \code{perRow}: Applies cutoff per row (i.e. gene). Internally, \code{rowSums()}
values are checked against this cutoff threshold prior to the melt
operation.
}}

\item{trans}{\code{character(1)}.
Apply a log transformation (e.g. \code{log2(x + 1L)}) to the count matrix prior
to melting, if desired. Use \code{"identity"} to return unmodified (default).}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Unpivot column data from wide format to long format.
}
\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
dim(object)
df <- melt(object)
nrow(df)
print(df)
}
\seealso{
\itemize{
\item https://seananderson.ca/2013/10/19/reshape/
\item \code{reshape2::melt} (deprecated).
\item \code{tidyr::pivot_longer}.
\item Python \code{pandas.melt}.
}
}
