% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\name{metrics}
\alias{metrics}
\alias{metricsPerSample}
\alias{metrics,SingleCellExperiment-method}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

metricsPerSample(object, ...)

\S4method{metrics}{SingleCellExperiment}(object, return = c("tbl_df", "DataFrame"))

\S4method{metricsPerSample}{SingleCellExperiment}(
  object,
  fun = c("mean", "median", "sum"),
  return = c("tbl_df", "DataFrame")
)
}
\arguments{
\item{object}{Object.}

\item{return}{\code{character(1)}.
Return type. Uses \code{\link[base:match.arg]{match.arg()}} internally and defaults
to the first argument in the \code{character} vector.}

\item{fun}{\code{function}.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame}
}
\description{
Quality control metrics
}
\section{Functions}{
\itemize{
\item \code{metrics,SingleCellExperiment-method}: Cell-level metrics.

\item \code{metricsPerSample,SingleCellExperiment-method}: Sample-level metrics.
}}

\note{
Updated 2021-02-05.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- calculateMetrics(object)
x <- metrics(object)
print(x)
x <- metricsPerSample(object, fun = "mean")
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
