% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/normalize-methods.R
\name{normalize}
\alias{normalize}
\alias{normalize,SingleCellExperiment-method}
\title{Normalize expression using pre-computed size factors}
\usage{
normalize(object, ...)

\S4method{normalize}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
This function calculates log-normalized size-factor-adjusted counts that
are defined in \code{logcounts()} from the raw, non-normalized count matrix
defined in \code{counts()}.
}
\details{
If no library size factors are defined in \code{sizeFactors()}, these values will
be computed internally automatically using
\code{scuttle::computeLibraryFactors()}.
}
\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
object <- normalize(object)
head(BiocGenerics::sizeFactors(object))
SingleCellExperiment::logcounts(object)[seq_len(2L), seq_len(2L)]
}
\seealso{
\itemize{
\item \code{AcidExperiment::estimateSizeFactors()}.
\item \code{scuttle::computeLibraryFactors()}.
\item \code{scuttle::logNormCounts()}.
\item \code{scuttle::librarySizeFactors()}.
\item \code{scuttle::geometricSizeFactors()}.
\item \code{scuttle::medianSizeFactors()}.
\item \code{Seurat::NormalizeData()}.
\item \code{monocle3::preprocess_cds()}.
\item \code{monocle3::normalized_counts()}.
\item \code{SingleCellExperiment::normcounts()}.
\item \code{SingleCellExperiment::logcounts()}.
}
}
