% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sampleData-methods.R
\name{sampleData}
\alias{sampleData}
\alias{sampleData<-}
\alias{sampleData,SingleCellExperiment-method}
\alias{sampleData<-,SingleCellExperiment,DFrame-method}
\title{Sample data}
\usage{
sampleData(object, ...)

sampleData(object, ...) <- value

\S4method{sampleData}{SingleCellExperiment}(
  object,
  clean = TRUE,
  ignoreCols = c("^description$", "^genomeBuild$", "^qualityFormat$", "^samRef$"),
  denylistCols = c("^ident$", "^g2mScore$", "^sScore$", "^phase$", "^oldIdent$",
    "^origIdent$", "^res[0-9]+")
)

\S4method{sampleData}{SingleCellExperiment,DFrame}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{clean}{\code{logical(1)}.
Only return \code{factor} columns. Useful when working with objects that contain
quality control metrics in \code{colData()}.
For example, \code{bcbioRNASeq} and \code{DESeqDataSet} objects often contain
additional columns that aren't informative sample metadata.}

\item{ignoreCols}{\code{character} or \code{NULL}.
Only applies when \code{clean = TRUE}. Additional factor columns defined in
\code{colData} to be ignored as sample-level metadata. Particularly useful for
\code{SingleCellExperiment} objects, where cell-to-sample mappings are defined
using the \code{sampleId} column.}

\item{denylistCols}{\code{character} or \code{NULL}.
Column names that should not be treated as sample-level metadata.
Currently applicable only to \code{SingleCellExperiment} objects, which have
cell-level columns that can be difficult to distinguish, especially when
processed using Seurat, scater, etc.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Metadata that describes the samples.
}
\details{
This is a complement to the standard
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} function, but improves support
for accessing sample metadata for datasets where multiple items in the
columns map to a single sample (e.g. cells for a single-cell RNA-seq
experiment).
}
\note{
Updated 2023-03-27.
}
\section{SingleCellExperiment}{


Recommended \code{colData}:
\itemize{
\item \code{sampleId}: \code{factor} defining cell-to-sample mappings. These mappings
should use syntactically valid names. Note that this is not currently
required as we're supporting \code{SingleCellExperiment} objects from 1 sample,
but it's required for working with multiple samples in a single object.
}
}

\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
sampleData(object)

## Assignment support.
sampleData(object)[["batch"]] <- 1L
## `batch` column should be now defined.
sampleData(object)
}
