% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/zerosVsDepth-methods.R
\name{zerosVsDepth}
\alias{zerosVsDepth}
\alias{zerosVsDepth,SingleCellExperiment-method}
\alias{zerosVsDepth,SummarizedExperiment-method}
\alias{zerosVsDepth,matrix-method}
\alias{zerosVsDepth,sparseMatrix-method}
\title{Percentage of zeros vs. library depth}
\usage{
zerosVsDepth(object, ...)

\S4method{zerosVsDepth}{SingleCellExperiment}(object, assay = 1L)

\S4method{zerosVsDepth}{SummarizedExperiment}(object, assay = 1L)

\S4method{zerosVsDepth}{matrix}(object)

\S4method{zerosVsDepth}{sparseMatrix}(object)
}
\arguments{
\item{object}{Object.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
\code{matrix}.
}
\description{
Calculate and visualize the dropout rate.
}
\note{
Updated 2022-10-24.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
df <- zerosVsDepth(object)
summary(df)
colnames(df)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
