% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{AcidSingleCellTestsURL}
\alias{AcidSingleCellTestsURL}
\title{AcidSingleCell test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AcidSingleCellTestsURL
}
\description{
AcidSingleCell test data URL
}
\note{
Updated 2021-09-13.
}
\examples{
AcidSingleCellTestsURL
}
\keyword{internal}
