% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateCols-methods.R
\name{aggregateCols}
\alias{aggregateCols}
\alias{aggregateCols,SingleCellExperiment-method}
\title{Aggregate columns}
\usage{
aggregateCols(x, ...)

\S4method{aggregateCols}{SingleCellExperiment}(x, col = "aggregate", fun = "sum")
}
\arguments{
\item{x}{Object.}

\item{col}{\code{character(1)}.
Name of column in either \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} that defines the desired
aggregation groupings.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{\link[base:match.arg]{match.arg()}} internally.}

\item{...}{Additional arguments.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Aggregate columns
}
\note{
Updated 2021-02-13.
}
\section{Methods (by class)}{


\itemize{
\item \code{matrix}, \code{Matrix}:
Aggregate using a grouping \code{factor}.
\item \code{SummarizedExperiment}:
Aggregate data slotted in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{assays()}} using an automatically
generated grouping \code{factor}, which is obtained from a user-defined column
(\code{col} argument) in either the
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData()}} or
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} of the object. Slot an
\code{aggregate} column into \code{\link[SummarizedExperiment:SummarizedExperiment-class]{rowData()}}
for \code{\link[AcidExperiment:aggregateRows]{aggregateRows()}}, or into
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}} for \code{\link[AcidExperiment:aggregateCols]{aggregateCols()}}. This
method will define the \code{groupings} automatically, and perform the
aggregation.
}

}

\examples{
data(SingleCellExperiment_lanesplit, package = "AcidTest")

## SingleCellExperiment ====
x <- SingleCellExperiment_lanesplit
levels(colData(x)[["aggregate"]])
x <- aggregateCols(x = x, col = "aggregate", fun = "sum")
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
