% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,SingleCellExperiment,SingleCellExperiment-method}
\title{Combining or merging different Bioconductor data structures}
\usage{
combine(x, y, ...)

\S4method{combine}{SingleCellExperiment,SingleCellExperiment}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Combining or merging different Bioconductor data structures
}
\note{
Updated 2021-02-05.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")
str_pad <- stringr::str_pad

## SingleCellExperiment ====
x <- SingleCellExperiment
colnames(x) <- paste0(
    "cell",
    str_pad(
        string = seq_len(ncol(x)),
        width = 4L,
        pad = "0"
    )
)

y <- x
colnames(y) <- paste0(
    "cell",
    str_pad(
        string = seq_len(ncol(y)) + ncol(y),
        width = 4L,
        pad = "0"
    )
)

## Combine the two objects.
c <- combine(x, y)
sampleData(c)
print(c)
}
\seealso{
\itemize{
\item \code{help("merge.Matrix", "Matrix.utils")}.
}
}
