% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertSampleIdsToNames-methods.R
\name{convertSampleIdsToNames}
\alias{convertSampleIdsToNames}
\alias{convertSampleIdsToNames,SingleCellExperiment-method}
\title{Convert sample identifiers to names}
\usage{
convertSampleIdsToNames(object, ...)

\S4method{convertSampleIdsToNames}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with valid sample IDs remapped to
human-readable sample names.
}
\description{
Convert sample identifiers to names
}
\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
convertSampleIdsToNames(object)
}
