% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cpm-methods.R
\name{cpm}
\alias{cpm}
\alias{cpm,SingleCellExperiment-method}
\title{Counts per million}
\usage{
cpm(object, ...)

\S4method{cpm}{SingleCellExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}
}
\description{
Counts per million
}
\note{
Updated 2021-09-14.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment
object <- estimateSizeFactors(object)
cpm <- cpm(object)
class(cpm)
mean(cpm)
}
\seealso{
\itemize{
\item \code{SingleCellExperiment::cpm()}.
\item \code{edgeR::cpm()}.
\item \code{scater::calculateCPM()}.
}
}
\keyword{internal}
