% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSingleCellExperiment.R
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\title{Make a SingleCellExperiment object}
\usage{
makeSingleCellExperiment(..., reducedDims = S4Vectors::SimpleList())
}
\arguments{
\item{...}{Passthrough arguments to \code{makeSummarizedExperiment()}.}

\item{reducedDims}{\code{SimpleList}.
List containing matrices of cell coordinates in reduced space.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Make a SingleCellExperiment object
}
\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SimpleList ====
object <- SingleCellExperiment_splatter
args <- list(
    "assays" = SummarizedExperiment::assays(object),
    "rowRanges" = SummarizedExperiment::rowRanges(object),
    "colData" = SummarizedExperiment::colData(object),
    "metadata" = S4Vectors::metadata(object),
    "reducedDims" = SingleCellExperiment::reducedDims(object)
)
sce <- do.call(what = makeSingleCellExperiment, args = args)
print(sce)
}
\seealso{
\itemize{
\item \code{AcidExperiment::makeSummarizedExperiment()}.
\item \code{SingleCellExperiment::altExps()}, which has replaced the now defunct
\code{isSpike} method for setting spike-in transcripts.
}
}
