% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSingleCellExperiment.R
\name{makeSingleCellExperiment}
\alias{makeSingleCellExperiment}
\title{Make a SingleCellExperiment object}
\usage{
makeSingleCellExperiment(..., reducedDims = SimpleList())
}
\arguments{
\item{...}{Passthrough arguments to
\code{\link[AcidExperiment:makeSummarizedExperiment]{makeSummarizedExperiment()}}.}

\item{reducedDims}{\code{SimpleList}.
List containing matrices of cell coordinates in reduced space.}
}
\value{
\code{SingleCellExperiment}.
}
\description{
Make a SingleCellExperiment object
}
\note{
Updated 2021-02-05.
}
\examples{
data(SingleCellExperiment, package = "AcidTest")

## SimpleList ====
object <- SingleCellExperiment
assays <- assays(object)
rowRanges <- rowRanges(object)
colData <- colData(object)
metadata <- metadata(object)
reducedDims <- reducedDims(object)

x <- makeSingleCellExperiment(
    assays = assays,
    rowRanges = rowRanges,
    colData = colData,
    metadata = metadata,
    reducedDims = reducedDims
)
print(x)
}
\seealso{
\itemize{
\item \code{AcidExperiment::makeSummarizedExperiment()}.
\item \code{SingleCellExperiment::altExps()}, which has replaced the now defunct
\code{isSpike} method for setting spike-in transcripts.
}
}
