% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\name{metrics}
\alias{metrics}
\alias{metricsPerSample}
\alias{metrics,SingleCellExperiment-method}
\alias{metricsPerSample,SingleCellExperiment-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

metricsPerSample(object, ...)

\S4method{metrics}{SingleCellExperiment}(object)

\S4method{metricsPerSample}{SingleCellExperiment}(object, fun = c("mean", "median", "sum"))
}
\arguments{
\item{object}{Object.}

\item{fun}{\code{function}.}

\item{...}{Additional arguments.}
}
\value{
\code{data.frame}
}
\description{
Quality control metrics
}
\section{Functions}{
\itemize{
\item \code{metrics,SingleCellExperiment-method}: Cell-level metrics.

\item \code{metricsPerSample,SingleCellExperiment-method}: Sample-level metrics.
}}

\note{
Updated 2022-03-02.
}
\examples{
data(SingleCellExperiment_splatter, package = "AcidTest")

## SingleCellExperiment ====
object <- SingleCellExperiment_splatter
object <- AcidExperiment::calculateMetrics(object)
x <- metrics(object)
print(x)
x <- metricsPerSample(object, fun = "mean")
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
