% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{SimpleList}
\alias{SingleCellExperiment}
\alias{assays}
\alias{basenameSansExt}
\alias{calculateMetrics}
\alias{colData}
\alias{isInstalled}
\alias{logcounts}
\alias{logcounts<-}
\alias{metadata}
\alias{normcounts}
\alias{normcounts<-}
\alias{reducedDim}
\alias{reducedDim<-}
\alias{reducedDimNames}
\alias{reducedDimNames<-}
\alias{reducedDims}
\alias{reducedDims<-}
\alias{rowRanges}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sentenceCase}
\alias{SingleCellExperiment-class}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{AcidBase}{\code{\link[AcidBase]{basenameSansExt}}}

  \item{AcidExperiment}{\code{\link[AcidExperiment:reexports]{assays}}, \code{\link[AcidExperiment]{calculateMetrics}}, \code{\link[AcidExperiment:reexports]{colData}}, \code{\link[AcidExperiment:reexports]{rowRanges}}, \code{\link[AcidExperiment]{sampleNames}}, \code{\link[AcidExperiment:sampleNames]{sampleNames<-}}}

  \item{AcidGenerics}{\code{\link[AcidGenerics:reexports]{SimpleList}}, \code{\link[AcidGenerics:reexports]{metadata}}}

  \item{goalie}{\code{\link[goalie:check-vector-isInstalled]{isInstalled}}}

  \item{SingleCellExperiment}{\code{\link[SingleCellExperiment]{SingleCellExperiment}}, \code{\link[SingleCellExperiment:assays]{logcounts}}, \code{\link[SingleCellExperiment:assays]{logcounts<-}}, \code{\link[SingleCellExperiment:assays]{normcounts}}, \code{\link[SingleCellExperiment:assays]{normcounts<-}}, \code{\link[SingleCellExperiment:reducedDims]{reducedDim}}, \code{\link[SingleCellExperiment:reducedDims]{reducedDim<-}}, \code{\link[SingleCellExperiment:reducedDims]{reducedDimNames}}, \code{\link[SingleCellExperiment:reducedDims]{reducedDimNames<-}}, \code{\link[SingleCellExperiment]{reducedDims}}, \code{\link[SingleCellExperiment:reducedDims]{reducedDims<-}}}

  \item{syntactic}{\code{\link[syntactic]{sentenceCase}}}
}}

